requirements-lock.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
braceexpand==0.1.7
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.1
datasets==2.11.0
dill==0.3.6
docker-pycreds==0.4.0
einops==0.6.0
einops-exts==0.0.4
filelock==3.10.7
frozenlist==1.3.3
fsspec==2023.3.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.13.3
idna==3.4
inflection==0.5.1
Jinja2==3.1.2
joblib==1.2.0
lit==16.0.0
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.0
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
open-clip-torch==2.16.0
-e git+http://ddd:5NzjwKaLsafcdBVpMsjD@source.cyberspace.vn/thinhhnt/open_flamingo.git@98ae64e6a5dad10c65848fc24bea0df99ed9e409#egg=open_flamingo
packaging==23.0
pandas==1.5.3
pathtools==0.1.2
Pillow==9.4.0
protobuf==3.20.3
psutil==5.9.4
pyarrow==11.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
responses==0.18.0
scipy==1.10.1
sentencepiece==0.1.97
sentry-sdk==1.18.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sympy==1.11.1
timm==0.6.13
tokenizers==0.13.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers @ git+https://github.com/huggingface/transformers@c612628045822f909020f7eb6784c79700813eda
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
wandb==0.14.0
wcwidth==0.2.6
webdataset==0.2.48
xxhash==3.2.0
yarl==1.8.2
